Creacion del modulo de evaluacion y modifiacion de algunas vistas ya existentes

+++ app/src/main/java/com/example/udlearning/ui/screens/activity/SolveSessionScreen.kt
@@ -63,7 +63,7 @@ fun SolveSessionScreen(
                 }
             }
             viewModel.isFinished -> {
-                FinalScoreScreen(viewModel.score, viewModel.totalPossibleScore) {
+                FinalScoreScreen(viewModel.score.toInt(), viewModel.totalPossibleScore.toInt()) {
                     navController.popBackStack()
                 }
             }
